refactor: Profile 컴포넌트 분리
